+++ src/main/java/com/amt/dflipflop/Controllers/CheckoutController.java
@@ -1,15 +1,9 @@
-package com.example.dflipflop.Controllers;
+package com.amt.dflipflop.Controllers;
 
-import com.example.dflipflop.Entities.Product;
-import com.example.dflipflop.Repositories.ProductRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-
-import java.util.List;
 
 @Controller
 public class CheckoutController {
