+++ src/main/java/com/amt/dflipflop/Controllers/StoreController.java
@@ -1,7 +1,6 @@
-package com.example.dflipflop.Controllers;
-import com.example.dflipflop.Repositories.ProductRepository;
-import com.example.dflipflop.Entities.Product;
-import com.example.dflipflop.Services.ProductService;
+package com.amt.dflipflop.Controllers;
+import com.amt.dflipflop.Entities.Product;
+import com.amt.dflipflop.Services.ProductService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
