+++ src/main/java/com/amt/dflipflop/Repositories/ProductRepository.java
@@ -1,6 +1,6 @@
-package com.example.dflipflop.Repositories;
+package com.amt.dflipflop.Repositories;
 
-import com.example.dflipflop.Entities.Product;
+import com.amt.dflipflop.Entities.Product;
 import org.springframework.data.repository.CrudRepository;
 
 // This will be AUTO IMPLEMENTED by Spring into a Bean called productRepository
