+++ src/main/java/com/amt/dflipflop/WebSecurityConfig.java
@@ -1,4 +1,4 @@
-package com.example.dflipflop;
+package com.amt.dflipflop;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
