+++ src/test/java/com/amt/dflipflop/Controllers/HomeControllerTests.java
@@ -1,12 +1,10 @@
-package com.example.dflipflop.Controllers;
+package com.amt.dflipflop.Controllers;
 
-import static org.hamcrest.Matchers.containsString;
-import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
+import com.amt.dflipflop.Services.ProductService;
 import org.junit.jupiter.api.Test;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,14 +12,17 @@ import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.web.servlet.MockMvc;
 
-@WebMvcTest(CheckoutController.class)
-public class CheckoutControllerTests {
+@WebMvcTest(HomeController.class)
+public class HomeControllerTests {
 
     @Autowired
     private MockMvc mockMvc;
 
+    @MockBean
+    private ProductService productService;
+
     @Test
-    public void shouldDisplayCheckout() throws Exception {
-        this.mockMvc.perform(get("/checkout")).andDo(print()).andExpect(status().isOk());
+    public void shouldDisplayHome() throws Exception {
+        this.mockMvc.perform(get("/")).andDo(print()).andExpect(status().isOk());
     }
 }
