+++ src/test/java/com/amt/dflipflop/Controllers/StoreControllerTests.java
@@ -1,16 +1,11 @@
-package com.example.dflipflop.Controllers;
+package com.amt.dflipflop.Controllers;
 
-import static org.hamcrest.Matchers.containsString;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import com.example.dflipflop.Controllers.StoreController;
-import com.example.dflipflop.Entities.Product;
-import com.example.dflipflop.Services.ProductService;
+import com.amt.dflipflop.Services.ProductService;
 import org.junit.jupiter.api.Test;
 
 import org.springframework.beans.factory.annotation.Autowired;
