+++ src/test/java/com/amt/dflipflop/FlipFlopApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.dflipflop;
+package com.amt.dflipflop;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
